requirements.txt
streamlit
pandas
gsheetsdb
gspread
gspread_pandas
plotly.express
pathlib
